% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangeseller_functions.R
\docType{package}
\name{adexchangeseller_googleAuthR}
\alias{adexchangeseller_googleAuthR}
\alias{adexchangeseller_googleAuthR-package}
\title{Ad Exchange Seller API
Accesses the inventory of Ad Exchange seller users and generates reports.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 22:31:27
filename: /Users/mark/dev/R/autoGoogleAPI/googleadexchangesellerv1.auto/R/adexchangeseller_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/adexchange.seller
\item https://www.googleapis.com/auth/adexchange.seller.readonly
}
}

