% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangeseller_objects.R
\name{AdUnit}
\alias{AdUnit}
\title{AdUnit Object}
\usage{
AdUnit(code = NULL, id = NULL, name = NULL, status = NULL)
}
\arguments{
\item{code}{Identity code of this ad unit, not necessarily unique across ad clients}

\item{id}{Unique identifier of this ad unit}

\item{name}{Name of this ad unit}

\item{status}{Status of this ad unit}
}
\value{
AdUnit object
}
\description{
AdUnit Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

