% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangeseller_objects.R
\name{CustomChannel}
\alias{CustomChannel}
\title{CustomChannel Object}
\usage{
CustomChannel(CustomChannel.targetingInfo = NULL, code = NULL, id = NULL,
  name = NULL, targetingInfo = NULL)
}
\arguments{
\item{CustomChannel.targetingInfo}{The \link{CustomChannel.targetingInfo} object or list of objects}

\item{code}{Code of this custom channel, not necessarily unique across ad clients}

\item{id}{Unique identifier of this custom channel}

\item{name}{Name of this custom channel}

\item{targetingInfo}{The targeting information of this custom channel, if activated}
}
\value{
CustomChannel object
}
\description{
CustomChannel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other CustomChannel functions: \code{\link{CustomChannel.targetingInfo}}
}
