% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangeseller_objects.R
\name{Report.headers}
\alias{Report.headers}
\title{Report.headers Object}
\usage{
Report.headers()
}
\value{
Report.headers object
}
\description{
Report.headers Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Report functions: \code{\link{Report}}
}
